// WADNR.Web/src/app/shared/generated/model/project-update-expected-funding-step-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FundSourceAllocationRequestUpdateItemRequest } from './fund-source-allocation-request-update-item-request';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectUpdateExpectedFundingStepRequest { 
    EstimatedTotalCost?: number | null;
    ProjectFundingSourceNotes?: string | null;
    FundingSourceIDs?: Array<number> | null;
    AllocationRequests?: Array<FundSourceAllocationRequestUpdateItemRequest> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectUpdateExpectedFundingStepRequestForm { 
    EstimatedTotalCost?: FormControl<number>;
    ProjectFundingSourceNotes?: FormControl<string>;
    FundingSourceIDs?: FormControl<Array<number>>;
    AllocationRequests?: FormControl<Array<FundSourceAllocationRequestUpdateItemRequest>>;
}

export class ProjectUpdateExpectedFundingStepRequestFormControls { 
    public static EstimatedTotalCost = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectFundingSourceNotes = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FundingSourceIDs = (value: FormControlState<Array<number>> | Array<number> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<number>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static AllocationRequests = (value: FormControlState<Array<FundSourceAllocationRequestUpdateItemRequest>> | Array<FundSourceAllocationRequestUpdateItemRequest> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<FundSourceAllocationRequestUpdateItemRequest>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}