// WADNR.Web/src/app/shared/generated/model/funding-source-option.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class FundingSourceOption { 
    FundingSourceID?: number;
    FundingSourceName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface FundingSourceOptionForm { 
    FundingSourceID?: FormControl<number>;
    FundingSourceName?: FormControl<string>;
}

export class FundingSourceOptionFormControls { 
    public static FundingSourceID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FundingSourceName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}